beggining of frontend implementation
lots of filed removed and added in this commit,
pagination has been implemented using routes to have a single page
application.
on the home page added a calender to see the day in which a workout was
performed.
this calender can be clicked to see the workouts info and click a delete
button to delete them.